Backend opus lite perf opt

See merge request !105

+++ src/backend/opus/storage.py
@@ -339,17 +339,12 @@ class DBInterface(StorageIFace):
         else:
             raise UniqueIDException()
 
-    def find_and_del_rel(self, from_node, to_node, rel_type):
+    def find_and_del_rel(self, from_node, to_node):
         '''Finds a relation of type rel_type between two nodes
         and deletes it'''
-        rows = self.db.query("START from_node=node({from_id}),"
-                             " to_node=node({to_id}) "
-                             "MATCH from_node-[rel:" + rel_type + "]->to_node "
-                             " RETURN rel",
-                             from_id=from_node.id, to_id=to_node.id)
-        for row in rows:
-            rel = row['rel']
-            rel.delete()
+        for rel in from_node.relationships.outgoing:
+            if rel.end.id == to_node.id:
+                rel.delete()
 
     def delete_relationship(self, rel):
         '''Deletes relatioship given a relationship object'''
